[bcr-wdc-e2e-tests] Move test utils out of main
