+++ crates/bcr-wdc-e2e-tests/src/main.rs
@@ -3,20 +3,20 @@
 use bcr_wdc_webapi::keys::ActivateKeysetRequest;
 use bcr_wdc_webapi::quotes::EnquireReply;
 use bcr_wdc_webapi::quotes::{
-    BillInfo, EnquireRequest, IdentityPublicData, StatusReply, UpdateQuoteRequest,
-    UpdateQuoteResponse,
+    EnquireRequest, StatusReply, UpdateQuoteRequest, UpdateQuoteResponse,
 };
-use bitcoin::secp256k1::Keypair;
+
 use cashu::nuts::nut02 as cdk02;
-use cashu::{Amount, MintBolt11Request, MintBolt11Response};
-use rand::Rng;
+use cashu::{MintBolt11Request, MintBolt11Response};
 use tracing::info;
 use tracing_subscriber::filter::LevelFilter;
 // ----- local modules
 mod endpoints;
 mod rest_client;
+mod test_utils;
 use endpoints::*;
 use rest_client::RestClient;
+use test_utils::{generate_blinds, get_amounts, random_ebill};
 // ----- end imports
 
 #[derive(Debug, serde::Deserialize)]
@@ -30,74 +30,6 @@ fn setup_tracing() {
     tracing_subscriber::fmt()
         .with_max_level(LevelFilter::INFO)
         .init();
-}
-
-fn random_ebill() -> (Keypair, BillInfo, bitcoin::secp256k1::schnorr::Signature) {
-    let bill_id = bcr_wdc_webapi::test_utils::random_bill_id();
-    let (_, drawee) = bcr_wdc_webapi::test_utils::random_identity_public_data();
-    let (_, drawer) = bcr_wdc_webapi::test_utils::random_identity_public_data();
-    let (_, payee) = bcr_wdc_webapi::test_utils::random_identity_public_data();
-
-    let endorsees_size = rand::thread_rng().gen_range(0..3);
-    let mut endorsees: Vec<IdentityPublicData> = Vec::with_capacity(endorsees_size);
-
-    let (endorser_kp, endorser) = bcr_wdc_webapi::test_utils::random_identity_public_data();
-    endorsees.push(endorser);
-
-    let owner_key = bcr_wdc_utils::keys::test_utils::generate_random_keypair();
-
-    let amount = Amount::from(rand::thread_rng().gen_range(1000..100000));
-
-    let bill = BillInfo {
-        id: bill_id,
-        maturity_date: random_date(),
-        drawee,
-        drawer,
-        payee,
-        endorsees,
-        sum: amount.into(),
-    };
-
-    let signature = bcr_wdc_utils::keys::schnorr_sign_borsh_msg_with_key(&bill, &endorser_kp)
-        .expect("schnorr_sign_borsh_msg_with_key");
-
-    (owner_key, bill, signature)
-}
-
-fn random_date() -> String {
-    let start = chrono::Utc::now() + chrono::Duration::days(365);
-    let days = rand::thread_rng().gen_range(0..365);
-    (start + chrono::Duration::days(days)).to_rfc3339()
-}
-
-fn get_amounts(mut targ: u64) -> Vec<u64> {
-    // TODO see if there is an existing cashu implementation
-    let mut coins = Vec::new();
-    let mut bit_position = 0;
-    while targ > 0 {
-        if (targ & 1) == 1 {
-            coins.push(1 << bit_position);
-        }
-        targ >>= 1;
-        bit_position += 1;
-    }
-    coins
-}
-
-pub fn generate_blinds(
-    keyset_id: cdk02::Id,
-    amounts: &[Amount],
-) -> Vec<(
-    cashu::BlindedMessage,
-    cashu::secret::Secret,
-    cashu::SecretKey,
-)> {
-    let mut blinds = Vec::new();
-    for amount in amounts {
-        let blind = bcr_wdc_utils::keys::test_utils::generate_blind(keyset_id, *amount);
-        blinds.push(blind);
-    }
-    blinds
 }
 
 async fn can_mint_ebill(cfg: &MainConfig) {
